5/自動ノート生成/requirements.txt
opencv-python
pytest-shutil
pathlib
argparse
ffmpeg-python
SpeechRecognition
pyaudio
google-api-python-client
mecab
mecab-python3
unidic-lite
keybert
python-pptx